+++ CHANGES.rst
@@ -34,6 +34,9 @@ Fixes:
 - Bundle aggregation must use ++plone++static overrided versions if any.
   [ebrehault]
 
+- Fix bundle aggregation when bundle has no CSS (or no JS)
+  [ebrehault]
+
 
 5.1a1 (2016-03-31)
 ------------------